data-generator/requirements.txt
kafka-python~=2.0.2
urllib3~=2.2.1
pandas~=2.2.1
